requirements.txt
mcp>=1.0.0
python-dotenv>=1.0.0
pywebview>=5.0.0
pynput>=1.7.0
requests>=2.32.0
pytest>=9.0.0
